Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,7 +2,7 @@
 int main(){
     float n,m;
     scanf("%f %f",&n,&m);
-    int product=n*m;
+    float product=n*m;
     printf("Product: %f",&product);
     return 0;
 }
